requirements.txt
# requirements.txt
Pillow
open_clip_torch
torch
numpy
pandas
scikit-learn
tk